+++ frontend/src/scenes/insights/insightLogic.tsx
@@ -18,6 +18,7 @@ import { cleanFilters } from 'scenes/insights/utils/cleanFilters'
 import { dashboardsModel } from '~/models/dashboardsModel'
 import { pollFunnel } from 'scenes/funnels/funnelUtils'
 import { preflightLogic } from 'scenes/PreflightCheck/logic'
+import { extractObjectDiffKeys } from './utils'
 
 const IS_TEST_MODE = process.env.NODE_ENV === 'test'
 
@@ -549,48 +550,3 @@ export const insightLogic = kea<insightLogicType>({
         },
     }),
 })
-
-function extractObjectDiffKeys(oldObj: FilterType, newObj: FilterType, prefix: string = ''): Record<string, any> {
-    if (Object.keys(oldObj).length === 0) {
-        return []
-    }
-
-    let changedKeys: Record<string, any> = {}
-    for (const [key, value] of Object.entries(newObj)) {
-        // @ts-ignore
-        if (!objectsEqual(value, oldObj[key])) {
-            if (key === 'events') {
-                if (value.length !== oldObj.events?.length) {
-                    changedKeys['changed_events_length'] = oldObj.events?.length
-                } else {
-                    value.forEach((event: Record<string, any>, idx: number) => {
-                        // @ts-ignore
-                        const _k = extractObjectDiffKeys(event, oldObj[key][idx], `event_${idx}_`)
-                        changedKeys = {
-                            ...changedKeys,
-                            ..._k,
-                        }
-                    })
-                }
-            } else if (key === 'actions') {
-                if (value.length !== oldObj.actions?.length) {
-                    changedKeys['changed_actions_length'] = oldObj.actions?.length
-                } else {
-                    value.forEach((action: Record<string, any>, idx: number) => {
-                        // @ts-ignore
-                        const _k = extractObjectDiffKeys(action, oldObj[key][idx], `action_${idx}_`)
-                        changedKeys = {
-                            ...changedKeys,
-                            ..._k,
-                        }
-                    })
-                }
-            } else {
-                // @ts-ignore
-                changedKeys[`changed_${prefix}${key}`] = oldObj[key]
-            }
-        }
-    }
-
-    return changedKeys
-}
