+++ frontend/src/scenes/insights/utils.ts
@@ -1,5 +1,5 @@
 import { BuiltLogic, Logic } from 'kea'
-import { InsightLogicProps, InsightType, ViewType, EntityFilter, ActionFilter } from '~/types'
+import { InsightLogicProps, InsightType, ViewType, EntityFilter, ActionFilter, FilterType } from '~/types'
 import { funnelLogic } from 'scenes/funnels/funnelLogic'
 import { retentionTableLogic } from 'scenes/retention/retentionTableLogic'
 import { pathsLogic } from 'scenes/paths/pathsLogic'
@@ -31,4 +31,53 @@ export const getDisplayNameFromEntityFilter = (
 
     // Return custom name. If that doesn't exist then the name, then the id, then just null.
     return (isCustom ? customName : null) ?? name ?? (filter?.id ? `${filter?.id}` : null)
+}
+
+export function extractObjectDiffKeys(
+    oldObj: FilterType,
+    newObj: FilterType,
+    prefix: string = ''
+): Record<string, any> {
+    if (Object.keys(oldObj).length === 0) {
+        return []
+    }
+
+    let changedKeys: Record<string, any> = {}
+    for (const [key, value] of Object.entries(newObj)) {
+        // @ts-ignore
+        if (!objectsEqual(value, oldObj[key])) {
+            if (key === 'events') {
+                if (value.length !== oldObj.events?.length) {
+                    changedKeys['changed_events_length'] = oldObj.events?.length
+                } else {
+                    value.forEach((event: Record<string, any>, idx: number) => {
+                        // @ts-ignore
+                        const _k = extractObjectDiffKeys(event, oldObj[key][idx], `event_${idx}_`)
+                        changedKeys = {
+                            ...changedKeys,
+                            ..._k,
+                        }
+                    })
+                }
+            } else if (key === 'actions') {
+                if (value.length !== oldObj.actions?.length) {
+                    changedKeys['changed_actions_length'] = oldObj.actions?.length
+                } else {
+                    value.forEach((action: Record<string, any>, idx: number) => {
+                        // @ts-ignore
+                        const _k = extractObjectDiffKeys(action, oldObj[key][idx], `action_${idx}_`)
+                        changedKeys = {
+                            ...changedKeys,
+                            ..._k,
+                        }
+                    })
+                }
+            } else {
+                // @ts-ignore
+                changedKeys[`changed_${prefix}${key}`] = oldObj[key]
+            }
+        }
+    }
+
+    return changedKeys
 }
